+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added LeveledSkillTreeNode#description.
 - Fixed the levels of the nodes in Character#skillNodes was incorrect.
 - Changed type of LeveledSkillTreeNode#level to SkillLevel.
+- Added StatPropertyValue#nameSkillTree.
 # 0.5.0
 **This version includes Breaking Changes**
 - Renamed Character#skills to skillTreeNodes.